# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Cardano Transaction Submit API",
        "title": "tx-submit-api",
        "contact": {
            "name": "Blink Labs",
            "url": "https://blinklabs.io",
            "email": "support@blinklabs.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v0"
    },
    "basePath": "/",
    "paths": {
        "/api/hastx/{tx_hash}": {
            "get": {
                "description": "Determine if a given transaction ID exists in the node mempool.",
                "produces": [
                    "application/json"
                ],
                "summary": "HasTx",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction Hash",
                        "name": "tx_hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/submit/tx": {
            "post": {
                "description": "Submit an already serialized transaction to the network.",
                "produces": [
                    "application/json"
                ],
                "summary": "Submit Tx",
                "parameters": [
                    {
                        "enum": [
                            "application/cbor"
                        ],
                        "type": "string",
                        "description": "Content type",
                        "name": "Content-Type",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}